src/domain/forum/application/use-cases/edit-answer/edit-answer.ts
import { left, right } from "@core/either";
import { AnswersRepository } from "@forum-repositories/answers-repository";
import { NotAllowedError } from "@forum-use-case-errors/not-allowed-error";
import { ResourceNotFoundError } from "@forum-use-case-errors/resource-not-found-error";
import {
	EditAnswerUseCaseProps,
	EditAnswerUseCaseResponse,
} from "./edit-answer.types";

export class EditAnswerUseCase {
	constructor(private answersRepository: AnswersRepository) {}

	async execute({
		authorId,
		answerId,
		content,
	}: EditAnswerUseCaseProps): Promise<EditAnswerUseCaseResponse> {
		const answer = await this.answersRepository.findById(answerId);

		if (!answer) {
			return left(new ResourceNotFoundError());
		}

		if (authorId !== answer.authorId.toString()) {
			return left(new NotAllowedError());
		}

		answer.content = content;

		await this.answersRepository.save(answer);

		return right({
			answer,
		});
	}
}
